CLK32-920: Update View
